cosmetic: move asynccache to its own package
